days 7 and 8

+++ AdventOfCode2024/Solvers/Day1/Day1Solver.cs
@@ -23,14 +23,14 @@ internal class Day1Solver : SolverBase<LocationIdLists>
         };
     }
 
-    protected override int SolvePart1(LocationIdLists input)
+    protected override ulong SolvePart1(LocationIdLists input)
     {
-        return input.LeftList.Order().Zip(input.RightList.Order()).Select((pair) => Math.Abs(pair.First - pair.Second)).Sum();
+        return (ulong)input.LeftList.Order().Zip(input.RightList.Order()).Select((pair) => Math.Abs(pair.First - pair.Second)).Sum();
     }
 
-    protected override int SolvePart2(LocationIdLists input)
+    protected override ulong SolvePart2(LocationIdLists input)
     {
         var rightCounts = input.RightList.GroupBy(x => x).ToDictionary(x => x.Key, x => x.Count());
-        return input.LeftList.Select(l => l * (rightCounts.TryGetValue(l, out int value) ? value : 0)).Sum();
+        return (ulong)input.LeftList.Select(l => l * (rightCounts.TryGetValue(l, out int value) ? value : 0)).Sum();
     }
 }
