+++ AdventOfCode2024/Solvers/Interface/DayResult.cs
@@ -2,7 +2,7 @@ namespace AdventOfCode2024.Solvers.Interface;
 
 public sealed record DayResult
 {
-    public required int SolutionPart1 { get; init; }
+    public required ulong SolutionPart1 { get; init; }
 
-    public required int SolutionPart2 { get; init; }
+    public required ulong SolutionPart2 { get; init; }
 }
